(#387): Moving the GPAP protocol stuff into its own library
